backend/python/pyproject.toml
[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "W", "I", "F401", "PLW0603"]
ignore = [
    "E501",  # Line too long
]
exclude = [
    "../../tests",
]
src = ["."]


[project]
name = "ai-service"
authors = [
    {name = "Abhishek", email = "abhishek@pipeshub.com"}
]
description = "AI Pipeline Service"
version = "0.1"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "aioboto3>=12.0.0",
    "aiohttp==3.11.2",
    "aiokafka==0.12.0",
    "aiolimiter==1.2.1",
    "aioredis==2.0.1",
    "azure-ai-formrecognizer==3.3.3",
    "box-sdk-gen==1.12.0",
    "celery==5.4.0",
    "confluent-kafka==2.8.0",
    "cryptography==44.0.1",
    "dependency-injector==4.44.0",
    "discord.py==2.6.3",
    "docling==2.53.0",
    "dropbox==12.0.2",
    "docx2python==3.5.0",
    "etcd3==0.12.0",
    "evernote3>=1.25.14",
    "fastapi==0.115.6",
    "fastembed==0.5.1",
    "html-to-markdown==2.3.0",
    "PyGithub==2.8.1",
    "python-gitlab==6.4.0",
    "google-api-python-client==2.161.0",
    "google-auth-oauthlib==1.2.1",
    "Jinja2==3.1.6",
    "jsonschema==4.23.0",
    "langchain-anthropic==0.3.17",
    "langchain-aws==0.2.18",
    "langchain-cohere==0.4.5",
    "langchain-community==0.3.18",
    "langchain-experimental==0.3.4",
    "langchain-fireworks==0.3.0",
    "langchain-google-genai==2.1.8",
    "langchain-google-vertexai==2.0.18",
    "langchain-groq==0.3.6",
    "langchain-huggingface==0.3.0",
    "langchain-mistralai==0.2.11",
    "langchain-ollama==0.3.0",
    "langchain-openai==0.3.28",
    "langchain-qdrant==0.2.0",
    "langchain-together==0.3.0",
    "langchain-voyageai==0.1.6",
    "langchain-xai==0.2.4",
    "langchain==0.3.19",
    "langgraph==0.3.34",
    "layoutparser==0.3.4",
    "linkedin-api-client>=0.3.0",
    "litellm==1.72.4",
    "markdown==3.8",
    "msgraph-sdk==1.16.0",
    "nats-py==2.1.0",
    "notion-client==2.2.1",
    "nltk==3.9.1",
    "numpy<2",
    "ocrmypdf==16.8.0",
    "openpyxl==3.1.5",
    "pandas==2.2.3",
    "pdf2image==1.17.0",
    "protobuf==3.20.3",
    "pydantic==2.10.5",
    "PyMuPDF==1.24.14",
    "python-arango==8.1.5",
    "python-docx==1.1.2",
    "python-dotenv==1.0.1",
    "python-jose==3.4.0",
    "python-multipart==0.0.18",
    "qdrant-client==1.13.1",
    "redis==5.2.1",
    "requests-mock==1.12.1",
    "requests-toolbelt==1.0.0",
    "requests==2.32.4",
    "ruff==0.11.9",
    "sentence-transformers==3.4.1",
    "slack-sdk==3.27.0",
    "spacy<=3.7.5",
    "tenacity==8.5.0",
    "thrift>=0.16.0",
    "uuid==1.30",
    "uvicorn==0.30.6",
]

[tool.setuptools.packages.find]
exclude = ["modules"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
module = "app.indexing_main:run"

[project.optional-dependencies]
dev = ["ruff"]